+++ src/constants.ts
@@ -10,36 +10,32 @@ export const DEFAULT_FONT_FAMILY =
   "-apple-system, 'Helvetica Neue', Arial, sans-serif";
 
 export const DTYPE_VALUES = {
-  '<u1': {
+  Uint8: {
     format: GL.R8UI,
     dataFormat: GL.RED_INTEGER,
     type: GL.UNSIGNED_BYTE,
-    max: 2 ** 8 - 1,
-    TypedArray: Uint8Array
+    max: 2 ** 8 - 1
   },
-  '<u2': {
+  Uint16: {
     format: GL.R16UI,
     dataFormat: GL.RED_INTEGER,
     type: GL.UNSIGNED_SHORT,
-    max: 2 ** 16 - 1,
-    TypedArray: Uint16Array
+    max: 2 ** 16 - 1
   },
-  '<u4': {
+  Uint32: {
     format: GL.R32UI,
     dataFormat: GL.RED_INTEGER,
     type: GL.UNSIGNED_INT,
-    max: 2 ** 32 - 1,
-    TypedArray: Uint32Array
+    max: 2 ** 32 - 1
   },
-  '<f4': {
+  Float32: {
     format: GL.R32F,
     dataFormat: GL.RED,
     type: GL.FLOAT,
     // Not sure what to do about this one - a good use case for channel stats, I suppose:
     // https://en.wikipedia.org/wiki/Single-precision_floating-point_format.
-    max: 3.4 * 10 ** 38,
-    TypedArray: Float32Array
+    max: 3.4 * 10 ** 38
   }
-};
+} as const;
 
-export const GLOBAL_SLIDER_DIMENSION_FIELDS = ['z', 'time'];
+export const GLOBAL_SLIDER_DIMENSION_FIELDS = ['z', 't'];