+++ src/loaders/decoder.js
@@ -1,5 +1,0 @@
-import { getDecoder } from 'geotiff/src/compression';
-
-export function getTiffDecoder(fileDirectory) {
-  return getDecoder(fileDirectory);
-}